requirements.txt
annotated-types==0.6.0
appnope==0.1.4
astroid==2.5
asttokens==2.4.1
atomicwrites==1.4.1
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.1
decorator==4.4.2
difPy==4.1.0
docstring_parser==0.16
executing==2.0.1
ExifRead==3.0.0
filelock==3.13.4
Flask==3.0.3
fonttools==4.51.0
fsspec==2024.3.1
geographiclib==2.0
geopy==2.4.1
google-ai-generativelanguage==0.6.1
google-api-core==2.18.0
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-aiplatform==1.47.0
google-cloud-bigquery==3.20.1
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.16.0
google-cloud-texttospeech==2.16.3
google-crc32c==1.5.0
google-generativeai==0.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpc-google-iam-v1==0.13.0
grpcio==1.62.1
grpcio-status==1.62.1
httplib2==0.22.0
huggingface-hub==0.22.2
idna==3.7
imagededup==0.3.1
ImageHash==4.0
imageio==2.34.0
imageio-ffmpeg==0.4.9
ipykernel==6.29.4
ipython==8.23.0
isort==4.3.21
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.0
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
lazy-object-proxy==1.10.0
lazy_loader==0.4
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.6
mccabe==0.6.1
more-itertools==10.2.0
moviepy==1.0.3
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
piexif==1.1.3
pillow==10.3.0
pillow_heif==0.16.0
platformdirs==4.2.0
pluggy==0.13.1
proglog==0.1.10
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.0
pydantic_core==2.18.1
pyexifinfo==0.4.0
Pygments==2.17.2
pyheif==0.7.1
pylint==2.5.0
pyparsing==3.1.2
pytest==4.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyWavelets==1.6.0
PyYAML==6.0.1
pyzmq==25.1.2
regex==2023.12.25
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
safetensors==0.4.2
scikit-image==0.23.1
scikit-learn==1.4.2
scipy==1.13.0
sentence-transformers==2.6.1
shapely==2.0.3
six==1.16.0
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
threadpoolctl==3.4.0
tifffile==2024.2.12
tokenizers==0.15.2
toml==0.10.2
torch==2.2.2
torchvision==0.17.2
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
transformers==4.39.3
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
versioneer==0.29
Wand==0.6.13
wcwidth==0.2.13
Werkzeug==3.0.2
wrapt==1.12.1
